backend/requirements.txt
requests
python-dotenv
Flask
Flask-Cors
gunicorn